third_party/WebKit/LayoutTests/hittesting/paint-containment-hittest-expected.txt

Hit testing should respect clips established by contain: paint.

PASS successfullyParsed is true

TEST COMPLETE
PASS document.elementFromPoint(centerX, centerY).id is 'containTransform'
PASS document.elementFromPoint(centerX, centerY).id is 'containAbsolute'
PASS document.elementFromPoint(centerX, centerY).id is 'containFixed'
PASS document.elementFromPoint(centerX, centerY).id is 'body'
PASS document.elementFromPoint(centerX, centerY).id is 'body'
PASS document.elementFromPoint(centerX, centerY).id is 'body'
